# metal : improve `MUL_MAT_ID`